Move transform operations to separate files
