+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/configs/interpolators/base/common.h
@@ -1,11 +1,15 @@
 #pragma once
 
 #include <reanimated/CSS/common/TransformMatrix3D.h>
-#include <reanimated/CSS/interpolation/InterpolatorFactory.h>
-#include <reanimated/CSS/interpolation/transforms/TransformOperation.h>
-
 #include <reanimated/CSS/configs/interpolators/constants.h>
 #include <reanimated/CSS/configs/interpolators/utils.h>
+#include <reanimated/CSS/interpolation/InterpolatorFactory.h>
+#include <reanimated/CSS/interpolation/transforms/operations/matrix.h>
+#include <reanimated/CSS/interpolation/transforms/operations/perspective.h>
+#include <reanimated/CSS/interpolation/transforms/operations/rotate.h>
+#include <reanimated/CSS/interpolation/transforms/operations/scale.h>
+#include <reanimated/CSS/interpolation/transforms/operations/skew.h>
+#include <reanimated/CSS/interpolation/transforms/operations/translate.h>
 
 #include <reanimated/CSS/common/values/CSSColor.h>
 #include <reanimated/CSS/common/values/CSSKeyword.h>
