+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/interpolation/transforms/TransformInterpolator.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <reanimated/CSS/interpolation/transforms/TransformOperation.h>
+#include <reanimated/CSS/interpolation/transforms/operations/TransformOperation.h>
 #include <reanimated/CSS/misc/ViewStylesRepository.h>
 
 #include <memory>
