+++ packages/react-native-reanimated/Common/cpp/reanimated/CSS/interpolation/transforms/operations/perspective.h
@@ -1,0 +1,17 @@
+#pragma once
+
+#include <reanimated/CSS/interpolation/transforms/operations/TransformOperation.h>
+
+namespace reanimated::css {
+
+struct PerspectiveOperation final : public TransformOperationBase<CSSDouble> {
+  using TransformOperationBase<CSSDouble>::TransformOperationBase;
+
+  explicit PerspectiveOperation(double value);
+
+  TransformOperationType type() const override;
+  folly::dynamic valueToDynamic() const override;
+  TransformMatrix3D toMatrix() const override;
+};
+
+} // namespace reanimated::css
